refactor: restructure codebase into vertical slices for main and renderer
