+++ src/main/infrastructure/copilot/__tests__/copilotDetector.test.ts
