+++ src/main/infrastructure/providers/IssueTrackerProvider.ts
