+++ src/renderer/features/settings/components/GeneralSettings.tsx
